nlopt (python tests to be fixed)
